+++ ccmpred/objfun/treecd/cext/treecd.h
@@ -5,10 +5,10 @@
 #include "cd.h"
 
 void mutate_along_tree(
-	int32_t *n_children,
+	uint64_t *n_children,
 	flt *branch_lengths,
 	flt *x,
-	uint32_t nvert,
+	uint64_t nvert,
 	uint8_t *seqs,
 	uint32_t ncol,
 	flt mutation_rate
